* added unit tests

+++ app/src/main/java/com/dempsey/example/marvelapp/business/AppMarvelBusiness.java
@@ -2,6 +2,7 @@ package com.dempsey.example.marvelapp.business;
 
 import android.app.Activity;
 import android.support.annotation.NonNull;
+import android.support.annotation.VisibleForTesting;
 import com.dempsey.example.marvelapp.data.dao.remotedao.AppMarvelLocalDao;
 import com.dempsey.example.marvelapp.data.dao.remotedao.AppMarvelRemoteDao;
 import com.dempsey.example.marvelapp.data.model.Comic;
@@ -12,11 +13,14 @@ import org.joda.time.DateTime;
 import org.joda.time.DateTimeZone;
 import rx.Observable;
 
+import static android.support.annotation.VisibleForTesting.PACKAGE_PRIVATE;
+
 public class AppMarvelBusiness implements MarvelBusiness {
 
   private final AppMarvelRemoteDao remoteDao;
   private final AppMarvelLocalDao localDao;
 
+  @VisibleForTesting (otherwise = PACKAGE_PRIVATE)
   AppMarvelBusiness(@NonNull final AppMarvelRemoteDao remoteDao, @NonNull final AppMarvelLocalDao localDao) {
     this.remoteDao = remoteDao;
     this.localDao = localDao;
@@ -33,7 +37,11 @@ public class AppMarvelBusiness implements MarvelBusiness {
 
   @Override
   public void storeComicsToInternalStorage(@NonNull Comic comics) {
-    deleteAllFromStorage();
+
+    if (retrieveComicsFromInternalStorage() != null) {
+      deleteAllFromStorage();
+    }
+
     localDao.storeToInternal(comics);
   }
